\name{lorede-package}
\alias{lorede-package}
\alias{lorede}
\docType{package}
\title{\packageTitle{lorede}}
\description{\packageDescription{lorede}}
\section{Package Content}{\packageIndices{lorede}}
\author{\packageAuthor{lorede}}
\section{Maintainer}{\packageMaintainer{lorede}}
\keyword{package}
